+++ packages/wise_feedback/lib/src/widgets/linear_feedback.dart
@@ -262,6 +262,7 @@ class _LinearFeedbackState extends State<LinearFeedback> {
         fields: widget.template.fields,
         showPriority: widget.showPriority,
         categories: widget.categories,
+        onClose: () => BetterFeedback.of(context).hide(),
         onSubmit: (description, {extras}) =>
             _submit(onSubmit, description, extras),
       ),